feat(proxy): added proxy server support to scully
