+++ package.json
@@ -16,7 +16,7 @@
     "scully:dev:watch": "tsc -w -p ./scully/tsconfig.scully.json",
     "scully:dev:compile": "tsc -p    ./scully/tsconfig.scully.json",
     "scully:dev:run": "nodemon ./dist/scully/scully.js  --ext js",
-    "scully:serve": "nodemon ./dist/scully/scully.js serve --ext js ",
+    "scully:serve": "nodemon --ext js ./dist/scully/scully.js serve --tds --proxy proxy.conf.json",
     "scully:dev:all": "ng build @scullyio/ng-lib && ng build && npm run scully:dev:compile && node ./dist/scully/scully",
     "scully:publish:patch": "cd ./scully && tsc -p ./tsconfig.scully.json && npm version patch && cp ./package.json ../dist/scully &&cp ./readme.md ../dist/scully && cd ../dist/scully && npm publish --access=public",
     "scully:publish:minor": "tsc -p ./tsconfig.scully.json && npm version minor",
@@ -93,6 +93,7 @@
     "marked": "^0.7.0",
     "ng-packagr": "^9.0.0-rc.1",
     "nodemon": "^1.19.4",
+    "osascript": "^1.2.0",
     "prettier": "^1.19.1",
     "pretty-quick": "^2.0.1",
     "protractor": "~5.4.2",
