+++ proxy.conf.json
@@ -1,0 +1,7 @@
+{
+  "/api": {
+    "target": "http://localhost:8200",
+    "secure": false,
+    "pathRewrite": {"^/api": ""}
+  }
+}
