+++ scully/utils/cli-options.ts
@@ -46,3 +46,10 @@ export const {tds} = yargs
   .boolean('tds')
   .default('tds', false)
   .describe('tds', 'start TestDataServer').argv;
+
+export const {proxyConfigFile} = yargs
+  .string('proxy')
+  .default('proxy', undefined)
+  .alias('proxy', 'proxyConfigFile')
+  .alias('proxy', 'proxyConfig')
+  .describe('proxy', 'Load proxy config from file').argv;
