+++ scully/utils/dataServer.ts
@@ -15,6 +15,7 @@ export async function startDataServer(ssl: boolean) {
     dataServer.use(function(req, res, next) {
       res.header('Access-Control-Allow-Origin', `${req.get('origin')}`);
       res.header('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept');
+      // console.log('DS',req.path)
       next();
     });
